rename typedef to make clear it is a property map
