+++ Mesh_3/include/CGAL/Mesh_3/experimental/Sizing_field_with_aabb_tree.h
@@ -17,7 +17,7 @@
 
 #include <CGAL/Profile_counter.h>
 #include <CGAL/Delaunay_triangulation_3.h>
-#include "Get_facet_patch_id.h"
+#include "Facet_patch_id_map.h"
 #include "Get_curve_index.h"
 #include <CGAL/Mesh_3/Protect_edges_sizing_field.h> // for weight_modifier
 
@@ -32,7 +32,7 @@
 template <typename GeomTraits, typename MeshDomain,
           typename Input_facets_AABB_tree = typename MeshDomain::AABB_tree,
           typename Get_curve_index_ = CGAL::Default,
-          typename Get_facet_patch_id_ = CGAL::Default
+          typename Facet_patch_id_map_ = CGAL::Default
           >
 struct Sizing_field_with_aabb_tree
 {
@@ -66,23 +66,23 @@ struct Sizing_field_with_aabb_tree
     CGAL::Mesh_3::Get_curve_index<typename MeshDomain::Curves_AABB_tree::Primitive>
     >::type Get_curve_index;
   typedef typename CGAL::Default::Get<
-    Get_facet_patch_id_,
+    Facet_patch_id_map_,
     CGAL::Mesh_3::Facet_patch_id_map<MeshDomain,
                                      typename Input_facets_AABB_tree::Primitive>
-    >::type Get_facet_patch_id;
+    >::type Facet_patch_id_map;
 
   Sizing_field_with_aabb_tree
   (typename Kernel_::FT d,
    const Input_facets_AABB_tree_& aabb_tree,
    const MeshDomain& domain,
    Get_curve_index get_curve_index = Get_curve_index(),
-   Get_facet_patch_id get_facet_patch_id = Get_facet_patch_id()
+   Facet_patch_id_map facet_patch_id_map = Facet_patch_id_map()
    )
     : d_(d), aabb_tree(aabb_tree),
       domain(domain),
       dt(),
       get_curve_index(get_curve_index),
-      get_facet_patch_id(get_facet_patch_id)
+      facet_patch_id_map(facet_patch_id_map)
   {
     {
       Corner_index maximal_corner_index = 0;
@@ -231,10 +231,10 @@ struct Sizing_field_with_aabb_tree
 
         CGAL::Mesh_3::Filtered_projection_traits<
           typename Input_facets_AABB_tree_::AABB_traits,
-          Get_facet_patch_id
+          Facet_patch_id_map
           > projection_traits(ids.begin(), ids.end(),
                               aabb_tree.traits(),
-                              get_facet_patch_id);
+                              facet_patch_id_map);
 
         aabb_tree.traversal(p, projection_traits);
 
@@ -258,7 +258,7 @@ struct Sizing_field_with_aabb_tree
                                "Ids are { ")
               % group(setprecision(17),result)
               % group(setprecision(17),p)
-              % CGAL::oformat(get(get_facet_patch_id,
+              % CGAL::oformat(get(facet_patch_id_map,
                                   projection_traits.closest_point_and_primitive().second))
               % group(setprecision(17),
                       projection_traits.closest_point_and_primitive().first);
@@ -294,10 +294,10 @@ struct Sizing_field_with_aabb_tree
         CGAL::Mesh_3::Filtered_projection_traits
           <
             typename Input_facets_AABB_tree_::AABB_traits
-          , Get_facet_patch_id
+          , Facet_patch_id_map
           > projection_traits(ids.begin(), ids.end(),
                               aabb_tree.traits(),
-                              get_facet_patch_id);
+                              facet_patch_id_map);
 
         aabb_tree.traversal(p, projection_traits);
 
@@ -308,7 +308,7 @@ struct Sizing_field_with_aabb_tree
           return result;
         }
 
-        CGAL_assertion(ids.count(get(get_facet_patch_id,
+        CGAL_assertion(ids.count(get(facet_patch_id_map,
                                      projection_traits.closest_point_and_primitive().second)) == 0);
 
         result =
@@ -327,7 +327,7 @@ struct Sizing_field_with_aabb_tree
                              "Closest face id: %4%\n"
                              "Ids are { ")
             % result % p % curve_id
-            % CGAL::oformat(get(get_facet_patch_id,
+            % CGAL::oformat(get(facet_patch_id_map,
                                 projection_traits.closest_point_and_primitive().second));
           for(Patch_index i : ids) {
             s << CGAL::oformat(i) << " ";
@@ -346,7 +346,7 @@ struct Sizing_field_with_aabb_tree
                              "Closest face id: %4%\n"
                              "Ids are { ")
             % result % p % curve_id
-            % CGAL::oformat(get(get_facet_patch_id,
+            % CGAL::oformat(get(facet_patch_id_map,
                                 projection_traits.closest_point_and_primitive().second));
           for(Patch_index i : ids) {
             s << CGAL::oformat(i) << " ";
@@ -528,7 +528,7 @@ private:
   Corners_incident_patches  corners_incident_patches;
   Corners_incident_curves   corners_incident_curves;
   Get_curve_index     get_curve_index;
-  Get_facet_patch_id  get_facet_patch_id;
+  Facet_patch_id_map  facet_patch_id_map;
 };
 
 #endif // CGAL_MESH_3_SIZING_FIELD_WITH_AABB_TREE_H
